Move `FPFHSignature33` to feature types header
