+++ apps/cloud_composer/include/pcl/apps/cloud_composer/items/fpfh_item.h
@@ -41,7 +41,7 @@
 #include <pcl/features/fpfh.h>
 #include <pcl/visualization/pcl_plotter.h>
 #include <pcl/visualization/qvtk_compatibility.h>
-#include <pcl/point_types.h>
+#include <pcl/feature_types.h>
 
 namespace pcl {
 namespace cloud_composer {
