+++ features/include/pcl/features/fpfh.h
@@ -41,7 +41,7 @@
 #pragma once
 
 #include <pcl/features/feature.h>
-#include <pcl/point_types.h>       // for pcl::FPFHSignature33
+#include <pcl/feature_types.h>       // for pcl::FPFHSignature33
 
 namespace pcl
 {
